查询权限数据

+++ config-example/example-item/interface-account/src/main/java/ren/steve/account/pojo/Account.java
@@ -31,7 +31,7 @@ public class Account implements Serializable {
   private String name;
   private String password;
   private String nickname;
-  private Long authority;
+  private Integer authority;
   private String phone;
   private String email;
   private String province;
